<feat>(ProblemService)
- findById(아이디로 조회)
- save(로컬저장)
- manual(문제 수동 업데이트)

+++ src/main/java/com/daun/word/domain/problem/domain/repository/ProblemRepository.java
@@ -11,5 +11,5 @@ public interface ProblemRepository {
 
     Problem save(final Problem request);
 
-    Optional<Problem> findById(final Long id);
+    Optional<Problem> findById(final Integer id);
 }
